+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - new `net-glb` module for Global External Load balancer
 - new `project-factory` module in [`examples/factories`](./examples/factories)
 - Module `project`: add missing Service Identity Accounts: artifactregistry, composer.
+- new data-solutions example: Cloud Storage to Bigquery with Cloud Dataflow with least privileges
 
 ## [12.0.0] - 2022-01-11